adding docker container build and tests for github actions
