+++ .github/workflows/main.yml
@@ -1,0 +1,25 @@
+name: CI
+
+on:
+  pull_request:
+    branches_ignore: []
+
+jobs:
+  formatting:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      
+      - name: Setup python environment
+        run: conda env create --quiet urlstechie
+
+      - name: Run tests
+        run: |
+          # activate conda env
+          export PATH="/usr/share/miniconda/bin:$PATH"
+          source activate urlstechie
+          python setup.py install
+          pytest -vs -x tests/test_check.py tests/test_fileproc.py tests/test_files tests/test_urlproc.py
+
+      - name: Build container image
+        run: docker build -t urlstechie .
