+++ tests/test_check.py
@@ -89,7 +89,7 @@ def test_script(config_fname, cleanup, print_all, force_pass, rcount, timeout):
     # Generate command
     cmd = ["urlchecker", "check", "--subfolder", "_project", "--file-types", file_types,
            "--white-listed-files", "conf.py", "--white-listed-urls", white_listed_urls,
-           "--white-listed_patterns", white_listed_patterns, "--retry-count", retry_count,
+           "--white-listed_patterns", white_listed_patterns, "--retry-count", rcount,
            "--timeout", timeout]
 
     # Add boolean arguments
