+++ tests/test_urlproc.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 import pytest
 from urlchecker.core.fileproc import collect_links_from_file
-from urlchecker.core.urlproc import check_response_status_code, check_urls, remove_empty
+from urlchecker.core.urlproc import check_response_status_code, check_urls
 
 
 @pytest.mark.parametrize('file', ["tests/test_files/sample_test_file.md",
@@ -11,14 +11,6 @@ def test_check_urls(file):
     """
     test check urls check function.
     """
+    check_results = {"failed": [], "passed": []}
     urls = collect_links_from_file(file)
-    check_urls(file, urls, check_results=[[],[]])
-
-
-def test_remove_empty():
-    """
-    test that empty urls are removed
-    """
-    urls = ["notempty", "notempty", "", None]
-    if len(remove_empty(urls)) != 2:
-        raise AssertionError
+    check_urls(file, urls, check_results=check_results)
